Adelantos working

+++ back/src/adelantos/adelantos.controller.ts
@@ -12,7 +12,7 @@ export class AdelantosController extends BaseController {
 
             const result = await dataSource.query(
                 `SELECT * From PersonalAdelanto ade 
-                WHERE (ade.PersonalAdelantoAprobado IN (NULL, 'S') OR ade.PersonalAdelantoAplicaEl= CONCAT(FORMAT(CONVERT(INT, @2), '00'),'/',@1))
+                WHERE ((ade.PersonalAdelantoAprobado IN ('S') OR ade.PersonalAdelantoAplicaEl= CONCAT(FORMAT(CONVERT(INT, @2), '00'),'/',@1)) OR ade.PersonalAdelantoAprobado IS NULL)
                 AND ade.PersonalId = @0`, [personalId, Año, Mes]
             )
             this.jsonRes(result, res)
@@ -27,24 +27,33 @@ export class AdelantosController extends BaseController {
         try {
             await queryRunner.connect()
             await queryRunner.startTransaction()
+
+            if (!personalId) throw new Error('Falta cargar la persona')
+            if (!monto) throw new Error('Falta cargar el monto')
+
             // Max Val
             //Hay ya un adelanto sin aprob
             // Si hay, lo reemplazas
             const adelantoId = (await queryRunner.query(`
-            SELECT MAX(ade.PersonalAdelantoId) as max FROM PersonalAdelanto ade WHERE ade.PersonalId = @0`, [personalId]))[0].max
-            const adelantoExistente = await dataSource.query(
+            SELECT MAX(ade.PersonalAdelantoId) as max FROM PersonalAdelanto ade WHERE ade.PersonalId = @0`, [personalId]))[0].max + 1
+            const adelantoExistente = await queryRunner.query(
                 `DELETE From PersonalAdelanto 
-                WHERE (PersonalAdelantoAprobado IN (NULL))
+                WHERE (PersonalAdelantoAprobado IS NULL)
                 AND PersonalId = @0`, [personalId]
             )
             const result = await queryRunner.query(
                 `INSERT INTO PersonalAdelanto(
-                    PersonalAdelantoId, PersonalId, PersonalAdelantoMonto, PersonalAdelantoFechaSolicitud, PersonalAdelantoAprobado, PersonalAdelantoFechaAprobacion, PersonalAdelantoCantidadCuotas, PersonalAdelantoAplicaEl, PersonalAdelantoLiquidoFinanzas, PersonalAdelantoUltimaLiquidacion, PersonalAdelantoCuotaUltNro, PersonalAdelantoMontoAutorizado, PersonalAdelantoJerarquicoId, PersonalAdelantoPuesto, PersonalAdelantoUsuarioId, PersonalAdelantoDia, PersonalAdelantoTiempo) VALUES(@0, @1, 
-                    @2, @3, @4, @5, @6, @7, @8, @9, @10, @11, @12, @13, @14, @15, @16)
-                `, [adelantoId, personalId, monto, (new Date()), null, null, null, 1, null, '', null, 0, null, ip, null, null])
+                    PersonalAdelantoId, PersonalId, PersonalAdelantoMonto, PersonalAdelantoFechaSolicitud, PersonalAdelantoAprobado, PersonalAdelantoFechaAprobacion, PersonalAdelantoCantidadCuotas, PersonalAdelantoAplicaEl, PersonalAdelantoLiquidoFinanzas, PersonalAdelantoUltimaLiquidacion, PersonalAdelantoCuotaUltNro, PersonalAdelantoMontoAutorizado, PersonalAdelantoJerarquicoId, PersonalAdelantoPuesto, PersonalAdelantoUsuarioId, PersonalAdelantoDia, PersonalAdelantoTiempo)
+                    VALUES(
+                    @0, @1, @2, @3, @4, @5, @6, @7, @8, @9, @10, @11, @12, @13, @14, @15, @16)
+                `, [adelantoId, personalId, monto, (new Date()), null, null, 1, null, null, '', null, 0, null, ip, null, null, null])
+            
             await queryRunner.commitTransaction()
+            this.jsonRes([], res)
         } catch (error) {
             await queryRunner.rollbackTransaction()
+            if (typeof error == 'string') {this.errRes(error, res, error, 409)}
+            else {this.errRes(error, res, 'Error accediendo a la base de datos', 409)}
         }
         finally {
             await queryRunner.release()
